RUST-2240 Move `serde_json` API behind a feature flag (#567)
