+++ Cargo.toml
@@ -48,6 +48,7 @@ serde_path_to_error = ["dep:serde_path_to_error"]
 # should be used in conjunction with chrono-0_4 or uuid-0_8.
 serde_with-3 = ["dep:serde_with", "dep:serde"]
 serde = ["dep:serde"]
+serde_json-1 = ["dep:serde_json"]
 
 [lib]
 name = "bson"
@@ -57,7 +58,7 @@ ahash = "0.8.0"
 chrono = { version = "0.4.15", features = ["std"], default-features = false, optional = true }
 rand = "0.9"
 serde = { version = "1.0", features = ["derive"], optional = true }
-serde_json = { version = "1.0", features = ["preserve_order"] }
+serde_json = { version = "1.0", features = ["preserve_order"], optional = true }
 indexmap = "2.1.0"
 hex = "0.4.2"
 base64 = "0.22.1"
@@ -84,6 +85,7 @@ pretty_assertions = "0.6.1"
 proptest = "1.0.0"
 serde_bytes = "0.11"
 serde_path_to_error = "0.1.16"
+serde_json = "1"
 chrono = { version = "0.4", features = ["serde", "clock", "std"], default-features = false }
 
 [package.metadata.docs.rs]
